Introduce Descriptor and change construction order
The context shouldn't be shared. The previous patch already addresses
this in the public API and this patch changes the internal handling.

We construct the operation-context now inside the QGpcOperation and
provide proper parent<>child handling between them. We teach our users
that the finished signal is _always_ emitted for the lifetime of an RPC,
but this is not true! During construction, there are checks which can
fail, which would return a null operation. To properly fix that we need
to have the Context available on construction already (due to
QGrpcOperation's API). This makes anyway the most sense logically.

This patch doesn't fix that, but is a preparation to change
that in the following patches. The same holds true for providing the
Descriptor and the channel. This will be required to provide the
Interceptors logic.

Change-Id: If6c1205ff0fbf3f4f54eb27b83b6c59ffe7df4e5
Reviewed-by: Alexey Edelev <semlanik@gmail.com>