Test mockserver: Add explicit Tag completion tracking
Add tag tracking functionality through the explicit EmptyGuard class.
Every tag requires a reference of this tracker, so that we can guarantee
that after leaving the scope of a testcase, all inserted Tags actually
finished.

I would like to have stack based Data management but unfortunately the
windows implementation on the CI for grpc 1.51 is using the context
after finishing, so we still have to delete our data with the Finished
operation.

Fixes: QTBUG-141780
Pick-to: 6.10 6.8
Change-Id: I1ed0d8e999a7693b85b3080acc29170a775de8d4
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>